Update devINA219.c -update int

+++ src/boot/ksdk1.1.0/devINA219.c
@@ -522,8 +522,8 @@ int16_t INA219_getBusVoltage_raw() {
  *  @brief  Gets the bus voltage in volts
  *  @return the bus voltage converted to volts
  */
-int16_t INA219_getBusVoltage_V() {
-  int16_t value = INA219_getBusVoltage_raw();
+int32_t INA219_getBusVoltage_V() {
+  int32_t value = INA219_getBusVoltage_raw();
   return value * 0.001; // from mV to V
 }
